# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DKMO Management Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - in: query
          name: search
          required: false
          schema: { type: string }
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [upcoming, ongoing, completed, cancelled]
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [recent, name, dateAsc, dateDesc]
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1 }
        - in: query
          name: pageSize
          required: false
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (Admin or Event Manager)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        eventDate:
          type: ["string", "null"]
          format: date-time
        location: { type: string }
        budget: { type: number }
        description: { type: string }
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
        taskCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - name
        - location
        - budget
        - description
        - status
        - createdAt
        - updatedAt
    EventInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        eventDate:
          type: ["string", "null"]
          format: date-time
        location: { type: string }
        budget: { type: number, minimum: 0 }
        description: { type: string }
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
      required: [name]
    EventList:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Event" }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
      required: [items, page, pageSize, total]
...